fix(contracts): repair matrix admission and CI ownership regressions
