+++ src/jacobian/math/matrices/values.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from collections.abc import Iterable, Mapping
 from fractions import Fraction
 from itertools import pairwise
-from typing import Any, Literal, Self, cast
+from typing import Annotated, Any, Literal, Self, cast
 
 from pydantic import Field, field_validator, model_validator
 from pydantic.json_schema import JsonSchemaValue
@@ -499,6 +499,8 @@ def integer_matrix_axis_schema(maximum_axis: int) -> JsonSchemaValue:
     """
 
     schema: dict[str, Any] = IntegerMatrix.model_json_schema()
+    for axis in ("row_count", "column_count"):
+        schema["properties"][axis]["maximum"] = maximum_axis
     entries = schema["properties"]["entries"]
     entries["maxItems"] = maximum_axis
     row_schema = entries.get("items")
@@ -804,7 +806,10 @@ class EmbeddedRealSimpleNumberFieldMatrix(StrictModel):
         return self
 
 
-ExactRealMatrix = RationalMatrix | EmbeddedRealSimpleNumberFieldMatrix
+ExactRealMatrix = Annotated[
+    RationalMatrix | EmbeddedRealSimpleNumberFieldMatrix,
+    Field(discriminator="domain", json_schema_extra={"required": ["domain"]}),
+]
 
 
 __all__ = [
