+++ tests/integration/test_sparse_nullspace.py
