+++ tests/math/combinatorics/test_exact_cover.py
@@ -468,62 +468,3 @@ def test_schema_publishes_the_exact_cover_contract() -> None:
     )
     assert "NO_COVER" in schema["description"]
     assert "UNKNOWN" in schema["description"]
-
-
-@pytest.mark.parametrize("count", [257, 688, 1100])
-def test_forced_cover_beyond_recursive_depth(count: int) -> None:
-    from jacobian.catalog.catalog import Catalog
-    from jacobian.dispatch import invoke_operation
-
-    items = [f"p{i:04}" for i in range(count)]
-    payload = {
-        "instance": {
-            "primary_items": items,
-            "secondary_items": [],
-            "rows": [
-                {"row_id": f"r{i:04}", "items": [item]} for i, item in enumerate(items)
-            ],
-        },
-        "search_node_limit": count + 1,
-    }
-    result = invoke_operation(
-        "combinatorics.generalized_exact_cover.find", payload, Catalog.open()
-    ).output
-    assert result["status"] == "FOUND"
-    assert result["selected_row_ids"] == [f"r{i:04}" for i in range(count)]
-    assert all(item["multiplicity"] == 1 for item in result["item_multiplicities"])
-    payload["search_node_limit"] = count
-    assert (
-        invoke_operation(
-            "combinatorics.generalized_exact_cover.find", payload, Catalog.open()
-        ).output["status"]
-        == "UNKNOWN"
-    )
-
-
-def test_exact_cover_prices_primary_mask_width_even_with_few_rows() -> None:
-    from jacobian.catalog.catalog import Catalog
-    from jacobian.catalog.models import OperationDomainValidationError
-    from jacobian.dispatch import invoke_operation
-
-    items = [f"p{i:04}" for i in range(4096)]
-    instance = {
-        "primary_items": items,
-        "secondary_items": [],
-        "rows": [{"row_id": "all", "items": items}],
-    }
-    catalog = Catalog.open()
-    assert (
-        invoke_operation(
-            "combinatorics.generalized_exact_cover.find",
-            {"instance": instance, "search_node_limit": 6250},
-            catalog,
-        ).output["status"]
-        == "FOUND"
-    )
-    with pytest.raises(OperationDomainValidationError, match="item-scan"):
-        invoke_operation(
-            "combinatorics.generalized_exact_cover.find",
-            {"instance": instance, "search_node_limit": 6251},
-            catalog,
-        )
